# [CARBONDATA-3721][CARBONDATA-3590] Optimize Bucket Table