--- a/integration/spark/src/main/scala/org/apache/spark/sql/execution/strategy/CarbonLateDecodeStrategy.scala
+++ b/integration/spark/src/main/scala/org/apache/spark/sql/execution/strategy/CarbonLateDecodeStrategy.scala
@@ -621,6 +621,7 @@ private[sql] class CarbonLateDecodeStrategy extends SparkStrategy {
         }
       }
       if (bucketColumns.size == cols.size) {
+        // use HashPartitioning will not shuffle
         HashPartitioning(bucketColumns, numBuckets)
       } else {
         UnknownPartitioning(0)