### Why is this PR needed?
 Support Bucket Table consistent with spark to improve the join performance by avoid shuffle for bucket column. The same time, fix bugs about load/compact query of bucket.
 
 ### What changes were proposed in this PR?
Support Bucket Table and consistent with spark to improve the join performance by avoid shuffle for bucket column. Fix bugs also.
1. For create table, ddl support both tblproperties and clustered by like hive.
2. For loading, fix some problems in loading when bucket column specified, make it clusterd into diff files based on bucket column.
3. For query, the hash impl should either keep the same for a given value or keep same with parquet table, so that the join result of diff bucket tables give correct result.  By the way, the hash impl is configurable.
4. For compaction, group the block files based on bucket id, the data should hash into diff carbondata files also, otherwise the query flow will group the files based on bucket num, all data compacted will com into 1 partition and the join result will mismatch, the performace will very slow.
5. For tests, add 19 test cases in TableBucketingTestCase
    
 ### Does this PR introduce any user interface change?
 - No

 ### Is any new testcase added?
 - Yes

 ### Performance

    3 million data each table with distinct 60000， 50 buckets join				

    select count(*) from
			(select t1.*, t2.*
			from t7 t1, t7_ t2
			where t1.name = t2.name) temp
  +---------+
  |count(1) |
  +---------+
  |190781696|
  +---------+
  		
1. When there are buckets in carbon, The first query is 30% higher and the second query is 50%-60% higher than no bucket table.
2. The carbon of no bucket is 30% faster than the parquet of no bucket
3. Carbon with bucket and parquet with bucket are slow for the first time and 10% faster for the second time, roughly equal.
4. Both buckets are faster than their own no bucket tables.

	format	    bucket 2nd cost	    bucket 1st cost	    no-bucket 2nd cost	no-bucket 1st cost
	carbon	        2126 ms	         3453 ms	    		        4952 ms	    			5092 ms
	parquet		2484 ms	    	 2770 ms 			6327 ms	    			8779 ms
	orc	        	2687 ms	   	  3765 ms	    		8678 ms	    			6567 ms