@ravipesala 2020-02-28
@Zhangshunyu Bucketing is already supported in Carbon.  I wonder why all this code is added again to support it.  If there are any issues if we are facing please put the testcases first which are not working or raise a jira.