@Zhangshunyu 2020-02-28
@ravipesala pls check all the new testcases added in TableBucketingTestCase and the comment i added in the pr desc.
1. all data stored into 1 file
2. join with parquet return wrong result
3. after compaction it will store into file of bucket id 0
4. new insert flow not work in
5. the others pls check testcases added